refactor[constants.ts]: Remove functions, enums, icons and interfaces from constants file
