+++ src/frontend/src/constants/enums.ts
@@ -1,0 +1,8 @@
+/**
+ * enum for the different types of nodes
+ * @enum
+ */
+export enum TypeModal {
+    TEXT = 1,
+    PROMPT = 2,
+  }
