+++ src/frontend/src/modals/ApiModal/index.tsx
@@ -43,16 +43,13 @@ import {
 } from "../../components/ui/tabs";
 import {
   EXPORT_CODE_DIALOG,
-  getCurlCode,
-  getPythonApiCode,
-  getPythonCode,
 } from "../../constants/constants";
 import { darkContext } from "../../contexts/darkContext";
 import { PopUpContext } from "../../contexts/popUpContext";
 import { TabsContext } from "../../contexts/tabsContext";
 import { FlowType } from "../../types/flow/index";
 import { buildTweaks } from "../../utils/reactflowUtils";
-import { classNames } from "../../utils/utils";
+import { classNames, getPythonApiCode, getCurlCode, getPythonCode, } from "../../utils/utils";
 export default function ApiModal({ flow }: { flow: FlowType }) {
   const [open, setOpen] = useState(true);
   const { dark } = useContext(darkContext);
